+++ NEWS.md
@@ -10,8 +10,9 @@
  * `futurize()` gained argument `eval`, which can be used to return
    the futurized expression instead of evaluating it.
 
- * Add support for `times()` of **foreach**. It requires special care
-   for passing future options, e.g. `futurize(seed = FALSE)`.
+ * Add support for `times()` and `%:%` of **foreach**, which require
+   special care when it comes to pass future options,
+   e.g. `futurize(seed = FALSE)`.
  
 
 # Version 0.0.2 (2025-05-23)